Make product save transition to desired route after save
By making save async, we can transition to the index route after the
content has been persisted.  That's likely the nicest scenario as it
shows the data is stored.  You could get away without the async and just
assume the data is stored.  That would be perfect if there'd be some
async store system (which I think we could easily have).