feat: Implement basic product creation and overview
* feat: Implement basic product creation WIP

* Make product save transition to desired route after save

By making save async, we can transition to the index route after the
content has been persisted.  That's likely the nicest scenario as it
shows the data is stored.  You could get away without the async and just
assume the data is stored.  That would be perfect if there'd be some
async store system (which I think we could easily have).

* Use form action for saving

This gives us common browser behavior, like saving when enter is
pressed.

* Fix product model

* Basic product listing on index

The product listing on index gives us some basic information of the
products we have found.  It looks horrid but it's good enough to debug.

* chore: Cleanup and present products nicely

Co-authored-by: Aad Versteden <madnificent@gmail.com>